% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphframes.R
\name{graph2graphframe}
\alias{graph2graphframe}
\title{Representation of BioPlex PPIs using a GraphFrames backend}
\usage{
graph2graphframe(gr, spark.con)
}
\arguments{
\item{gr}{an object of class \code{\linkS4class{graph}} storing the
BioPlex PPIs. Typically obtained via \code{\link{bioplex2graph}}.}

\item{spark.con}{Spark connection. Typically obtained via
\code{sparklyr::spark_connect}.}
}
\value{
An object of class \code{GraphFrame}.
}
\description{
Representation of BioPlex PPIs in a \code{GraphFrame} object
from the \code{graphframes} package.
}
\examples{

library(sparklyr)
library(graphframes)

# (1) Obtain the latest version of the 293T PPI network
bp.293t <- getBioPlex(cell.line = "293T", version = "3.0")

# (2) Turn the data into a graph 
bp.gr <- bioplex2graph(bp.293t)

# (3) Switch to a graphframes backend
sc <- spark_connect(master = "local", version = "3.0")
bp.gf <- graph2graphframe(bp.gr, sc) 

}
\references{
BioPlex: \url{https://bioplex.hms.harvard.edu/interactions.php}
}
\seealso{
\code{\link{bioplex2graph}}
}
