% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphframes.R
\name{graphframe2graph}
\alias{graphframe2graph}
\title{Convert a GraphFrames object to a graphNEL object}
\usage{
graphframe2graph(gf)
}
\arguments{
\item{gf}{An object of class \code{GraphFrame}.}
}
\value{
An object of class \code{\linkS4class{graph}}
}
\description{
Conversion of a GraphFrames object from the graphframes package
to a graphNEL object from the graph package.
}
\seealso{
\code{\link{graph2graphframe}}
}
