% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{joinNetworks}
\alias{joinNetworks}
\title{Join multiple networks around a target node ID}
\usage{
joinNetworks(id, network, intA.col = "SymbolA", intB.col = "SymbolB")
}
\arguments{
\item{id}{character. A gene or protein ID around which a joined network
should be constructed.}

\item{network}{list. Each element of the list is a \code{data.frame} representing
a network. That means each row of a \code{data.frame} represents an interaction.}

\item{intA.col}{character. Name of the column that stores the ID of the first
interactor in each \code{data.frame} of argument \code{network}.}

\item{intB.col}{character. Name of the column that stores the ID of the second
interactor in each \code{data.frame} of argument \code{network}.}
}
\value{
A \code{data.frame} containing the joined network.
}
\description{
This function joins multiple networks and resticts the joined
network to interactions involving a target node ID of interest
}
\examples{
 library(BioPlex)
 bp.293t <- BioPlex::getBioPlex(cell.line = "293T", version = "3.0")
 bp.hct <- BioPlex::getBioPlex(cell.line = "HCT116", version = "1.0")
 netl <- list(HEK = bp.293t, HCT = bp.hct)

 jnet <- joinNetworks(id = "CDH2", network = netl)

}
