% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb.R
\name{viewPDB}
\alias{viewPDB}
\title{View a protein structure from PDB}
\usage{
viewPDB(pdb.id)
}
\arguments{
\item{pdb.id}{character. A PDB ID in 4-letter PDB code or 6-letter
PDB-ID_Chain-ID code.}
}
\value{
An html widget that can be explored in the browser.
}
\description{
This function spins up a 3D viewer of a protein structure
}
\examples{
 viewPDB("6nmi")

}
